Refactoring and v0.3.2-2-alpha.2
1) call tests with Test suffix like they're called in ooni-probe

2) put stuff in reasonable directories rather than keeping everything
   in a single place to avoid doing refactoring later and then
   confusing people (also this is in line probably with how I'd like
   to restructure MK proper and fix layers inside it)

3) make sure deprecated stuff is marked as such
