+++ java/org/openobservatory/measurement_kit/common/LogCallback.java
@@ -2,8 +2,8 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-package org.openobservatory.measurement_kit;
+package org.openobservatory.measurement_kit.common;
 
-public interface EntryCallback {
-    public void callback(final String entry);
+public interface LogCallback {
+    public void callback(final long verbosity, final String message);
 }
