+++ java/org/openobservatory/measurement_kit/nettests/EntryCallback.java
@@ -2,8 +2,8 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-package org.openobservatory.measurement_kit;
+package org.openobservatory.measurement_kit.nettests;
 
-public interface LogCallback {
-    public void callback(final long verbosity, final String message);
+public interface EntryCallback {
+    public void callback(final String entry);
 }
