+++ java/org/openobservatory/measurement_kit/nettests/NdtTest.java
@@ -2,8 +2,11 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-package org.openobservatory.measurement_kit;
+package org.openobservatory.measurement_kit.nettests;
 
-public interface TestCompleteCallback {
-    public void callback();
+public class NdtTest extends OoniTestBase {
+
+    public NdtTest() {
+        super("ndt");
+    }
 }
