+++ scripts/run-swig
@@ -2,12 +2,13 @@
 set -e
 ROOTDIR=$(cd $(dirname $(dirname $0)) && pwd -P)
 
-swig -c++ -package org.openobservatory.measurement_kit                         \
+swig -c++ -package org.openobservatory.measurement_kit.swig                    \
         -java jni/ooni_test_wrapper.i
 
+install -d java/org/openobservatory/measurement_kit/swig
 for suffix in "" SwigJNI; do
     mv jni/OoniTestWrapper${suffix}.java                                       \
-            java/org/openobservatory/measurement_kit
+            java/org/openobservatory/measurement_kit/swig
 done
 rm jni/OoniTestWrapperSwig.java  # basically an empty file
 mv jni/ooni_test_wrapper_wrap.cxx jni/ooni_test_wrapper_wrap.cpp
